fixed api

+++ ptolemy/actor/lib/fmi/maHybrid/shared/sim_support.h
@@ -73,7 +73,7 @@ void loadFMU(FMU *fmu, const char *fmuFileName);
 #ifndef _MSC_VER
 typedef int boolean;
 #endif
-void outputRow(FMU fmus[], int numberOfFMUs, const char* NAMES_OF_FMUS[], int time, int resolution, FILE* file, char separator, boolean header);
+void outputRow(FMU fmus[], int numberOfFMUs, const char* NAMES_OF_FMUS[], double time, int resolution, FILE* file, char separator, boolean header);
 int error(const char *message);
 void printHelp(const char *fmusim);
 char *getTempResourcesLocation(); // caller has to free the result
